+++ examples/svm/plot_separating_hyperplane_unbalanced.py
@@ -25,7 +25,7 @@ X = np.r_[1.5 * np.random.randn(n_samples_1, 2),
 y = [0] * (n_samples_1) + [1] * (n_samples_2)
 
 # fit the model and get the separating hyperplane
-clf = svm.SVC(kernel='linear')
+clf = svm.SVC(kernel='linear', C=100)
 clf.fit(X, y)
 
 w = clf.coef_[0]
@@ -35,8 +35,8 @@ yy = a * xx - clf.intercept_[0] / w[1]
 
 
 # get the separating hyperplane using weighted classes
-wclf = svm.SVC(kernel='linear')
-wclf.fit(X, y, class_weight={1: 10})
+wclf = svm.SVC(kernel='linear', class_weight={1: 10})
+wclf.fit(X, y)
 
 ww = wclf.coef_[0]
 wa = -ww[0] / ww[1]
@@ -44,10 +44,10 @@ wyy = wa * xx - wclf.intercept_[0] / ww[1]
 
 # plot separating hyperplanes and samples
 pl.set_cmap(pl.cm.Paired)
-h0 = pl.plot(xx, yy, 'k-')
-h1 = pl.plot(xx, wyy, 'k--')
+h0 = pl.plot(xx, yy, 'k-', label='no weights')
+h1 = pl.plot(xx, wyy, 'k--', label='with weights')
 pl.scatter(X[:, 0], X[:, 1], c=y)
-pl.legend((h0, h1), ('no weights', 'with weights'))
+pl.legend()
 
 pl.axis('tight')
 pl.show()