+++ sklearn/linear_model/least_angle.py
@@ -16,7 +16,6 @@ from scipy.linalg.lapack import get_lapack_funcs
 
 from .base import LinearModel
 from ..utils import array2d, arrayfuncs, deprecated
-from ..utils.extmath import norm
 from ..cross_validation import check_cv
 from ..externals.joblib import Parallel, delayed
 
@@ -732,9 +731,6 @@ class LarsCV(LARS):
 
         Gram = 'auto' if self.precompute else None
 
-        # Scaling factor for the residues, to avoid overflows
-        X_norm = norm(X)
-
         cv_paths = Parallel(n_jobs=self.n_jobs, verbose=self.verbose)(
                     delayed(_lars_path_residues)(X[train], y[train],
                             X[test], y[test], Gram=Gram,
@@ -746,7 +742,8 @@ class LarsCV(LARS):
                             eps=self.eps)
                     for train, test in cv)
         all_alphas = np.concatenate(list(zip(*cv_paths))[0])
-        all_alphas.sort()
+        # Unique also sorts
+        all_alphas = np.unique(all_alphas)
         # Take at most max_n_alphas values
         stride = int(max(1, int(len(all_alphas) / float(self.max_n_alphas))))
         all_alphas = all_alphas[::stride]
@@ -764,7 +761,6 @@ class LarsCV(LARS):
             this_residues = interpolate.interp1d(alphas,
                                                  residues,
                                                  axis=0)(all_alphas)
-            this_residues /= X_norm
             this_residues **= 2
             mse_path[:, index] = np.mean(this_residues, axis=-1)